Course Ownership

+++ controllers/courses.js
@@ -48,11 +48,21 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @access    Private
 exports.addCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId;
+  req.body.user = req.user.id;
   const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
   if(!bootcamp){
     //res.status(400).json({success:false,message:`Course not found with id of ${req.params.id}`})
     return next(new ErrorResponse(`bootcamp not found with id of ${req.params.bootcampId}`, 404))
+  }
+  //Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
+        401
+      )
+    );
   }
 
   const course = await Course.create(req.body);
@@ -77,6 +87,15 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
     //res.status(400).json({success:false,message:`Course not found with id of ${req.params.id}`})
     return next(new ErrorResponse(`Course not found with id of ${req.params.id}`, 404))
   }
+  //Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update a course ${course._id}`,
+        401
+      )
+    );
+  }
 
   res.status(200).json({
       success:true,
@@ -94,6 +113,16 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
   if(!course){
     //res.status(400).json({success:false,message:`Course not found with id of ${req.params.id}`})
     return next(new ErrorResponse(`Course not found with id of ${req.params.id}`, 404))
+  }
+
+  //Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to delete the course ${course._id}`,
+        401
+      )
+    );
   }
 
   course = await course.deleteOne();
